refactor(agent): split Run and executeOne state machines / 拆分 Run 与 executeOne 状态机
Problem:
Agent.Run and executeOne mixed lifecycle, turn policy, and tool policy in
two very large methods, raising cognitive load without changing behavior.

Root cause:
Turn-loop counters and tool-call gating lived inline instead of package-
private staged helpers with a single ownership model for locks/defers.

Fix:
- Keep Run as the lifecycle owner of workspace lease, evidence commit,
  delivery checkpoint, steer queue, and active-turn cleanup.
- Extract runLoopState plus beginRunTurn, runToolLoop, handleFinalResponse,
  and handleToolRound without changing stream/persist/batch/compact order.
- Split executeOne into parseToolCall, resolveToolPolicy (plan/proxy,
  delivery, recovery/permission), prepareToolExecution, and
  finishToolExecution behind toolCallPlan, preserving permission-before-
  lease and PreToolUse-before-execute semantics.

Verification:
go test ./internal/agent/... -count=1
go test -race ./internal/agent/... -count=1
REASONIX_RELEASE_CACHE_GUARD=1 go test ./internal/agent -run '^TestReleaseCacheHitGuard$' -count=1
